+++ src/platform/forms-system/src/js/routing.js
@@ -1,6 +1,12 @@
 import _ from 'lodash/fp'; // eslint-disable-line no-restricted-imports
-import { getActiveExpandedPages } from './helpers';
+import {
+  getActiveExpandedPages,
+  createFormPageList,
+  createPageList,
+} from './helpers';
 
+import FormPage from './containers/FormPage';
+import ReviewPage from './review/ReviewPage';
 /*
  * Returns the page list without conditional pages that have not satisfied
  * their dependencies and therefore should be skipped.
@@ -25,4 +31,60 @@ export function getPreviousPagePath(pageList, data, pathname) {
   // if not, go back to the beginning because they shouldn’t be here
   const page = pageIndex >= 0 ? pageIndex - 1 : 0;
   return pages[page].path;
+}
+
+/*
+ * Create the routes based on a form config. This goes through each chapter in a form
+ * config, pulls out the config for each page, then generates a list of Route components with the
+ * config as props
+ */
+export function createRoutes(formConfig) {
+  const formPages = createFormPageList(formConfig);
+  const pageList = createPageList(formConfig, formPages);
+
+  let routes = formPages.map(page => ({
+    path: page.path,
+    component: page.component || FormPage,
+    pageConfig: page,
+    pageList,
+    urlPrefix: formConfig.urlPrefix,
+  }));
+
+  if (formConfig.additionalRoutes) {
+    routes = formConfig.additionalRoutes
+      .map(route => ({
+        ...route,
+        formConfig,
+        pageList,
+      }))
+      .concat(routes);
+  }
+
+  if (formConfig.introduction) {
+    routes = [
+      {
+        path: 'introduction',
+        component: formConfig.introduction,
+        formConfig,
+        pageList,
+      },
+    ].concat(routes);
+  }
+
+  return routes.concat([
+    {
+      path: 'review-and-submit',
+      formConfig,
+      component: ReviewPage,
+      pageList,
+    },
+    {
+      path: 'confirmation',
+      component: formConfig.confirmation,
+    },
+    {
+      path: '*',
+      onEnter: (nextState, replace) => replace(formConfig.urlPrefix || '/'),
+    },
+  ]);
 }
