+++ src/platform/forms-system/test/js/routing.unit.spec.js
@@ -1,6 +1,10 @@
 import { expect } from 'chai';
 
-import { getNextPagePath, getPreviousPagePath } from '../../src/js/routing';
+import {
+  getNextPagePath,
+  getPreviousPagePath,
+  createRoutes,
+} from '../../src/js/routing';
 
 describe('Schemaform routing', () => {
   function getPageList(dependsCallback) {
@@ -114,5 +118,44 @@ describe('Schemaform routing', () => {
 
     const path = getPreviousPagePath(pageList, data, pathname);
     expect(path).to.equal('/testing/0/conditional-page');
+  });
+  describe('createRoutes', () => {
+    it('should create routes', () => {
+      const formConfig = {
+        disableSave: true,
+        chapters: {
+          firstChapter: {
+            pages: {
+              testPage: {
+                path: 'test-page',
+              },
+            },
+          },
+        },
+      };
+
+      const routes = createRoutes(formConfig);
+
+      expect(routes[0].path).to.equal('test-page');
+      expect(routes[1].path).to.equal('review-and-submit');
+    });
+    it('should create routes with intro', () => {
+      const formConfig = {
+        introduction: f => f,
+        chapters: {
+          firstChapter: {
+            pages: {
+              testPage: {
+                path: 'test-page',
+              },
+            },
+          },
+        },
+      };
+
+      const routes = createRoutes(formConfig);
+
+      expect(routes[0].path).to.equal('introduction');
+    });
   });
 });
